gcp/requirements.txt
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.0
requests==2.31.0
flask==2.3.3
